Move import logic from ImportLabelsHandler to LabelImportPreProcessor
